[Refactor] Break okHttp impl into smaller files
